adpcm: rename codec_ima to codec_ima_wav
